# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TTS Storyteller API",
    "description": "Multi-backend TTS API supporting Qwen3-TTS and VibeVoice for voice-driven story generation with parallel execution",
    "version": "0.1.0"
  },
  "paths": {
...
    "/stories": {
      "get": {
        "summary": "List Stories Endpoint",
        "description": "List all stories.\n\nReturns story summaries with id, slug, and title.",
        "operationId": "list_stories_endpoint_stories_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/StorySummary"
                  },
                  "type": "array",
                  "title": "Response List Stories Endpoint Stories Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Story Endpoint",
        "description": "Create a new story template.\n\nVoice assignment: Use the `casting` field to map roleIds (as strings) to voiceIds.\nResolution order: line.actorId \u2192 casting[roleId] \u2192 defaultVoiceId\n\nExample casting: {\"0\": \"narrator_male\", \"1\": \"woman\"} assigns roleId 0 to narrator_male, roleId 1 to woman.\n\nReturns the story with server-generated `id` (UUID) and `slug` fields.",
        "operationId": "create_story_endpoint_stories_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StoryTemplate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoryTemplate"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/stories/{identifier}": {
      "get": {
        "summary": "Get Story Endpoint",
        "description": "Get story template by ID or slug.\n\nThe identifier can be either:\n- A UUID (e.g., \"550e8400-e29b-41d4-a716-446655440000\")\n- A slug (e.g., \"my_story_title\")",
        "operationId": "get_story_endpoint_stories__identifier__get",
        "parameters": [
          {
            "name": "identifier",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Identifier"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoryTemplate"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/stories/{storyId}": {
      "put": {
        "summary": "Replace Story Endpoint",
        "description": "Replace an existing story template.",
        "operationId": "replace_story_endpoint_stories__storyId__put",
        "parameters": [
          {
            "name": "storyId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Storyid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StoryTemplate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoryTemplate"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/stories/{storyId}/render": {
      "post": {
        "summary": "Render Story Endpoint",
        "description": "Resolve roles to voices for a story.\n\nReturns a preview of voice assignments showing which voice will be used for each line.\nResolution order: line.actorId \u2192 casting[roleId] \u2192 defaultVoiceId\n\nUse this endpoint to verify voice assignments before generating audio.",
        "operationId": "render_story_endpoint_stories__storyId__render_post",
        "parameters": [
          {
            "name": "storyId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Storyid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ResolvedLine"
                  },
                  "title": "Response Render Story Endpoint Stories  Storyid  Render Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/stories/{storyId}/generate": {
      "post": {
        "summary": "Generate Story Endpoint",
        "description": "Start audio generation for a story.\n\nOnly one job per story can be running or queued at a time.\nIf a job is already active for this story, returns 409 Conflict unless\ncancel_existing=true, in which case the active job is cancelled and a new one is started.",
        "operationId": "generate_story_endpoint_stories__storyId__generate_post",
        "parameters": [
          {
            "name": "storyId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Storyid"
            }
          },
          {
            "name": "cancel_existing",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Cancel Existing"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/GenerateRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Request"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "GenerateRequest": {
        "properties": {
          "concat": {
            "type": "boolean",
            "title": "Concat",
            "description": "Whether to concatenate outputs",
            "default": true
          }
        },
        "type": "object",
        "title": "GenerateRequest",
        "description": "Request body for story generation."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "Job": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "Job identifier"
          },
          "type": {
            "type": "string",
            "title": "Type",
            "description": "Job type (e.g., 'generate', 'voice_generate')"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "description": "Job status: queued, running, succeeded, failed"
          },
          "storyId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Storyid",
            "description": "Associated story ID"
          },
          "voiceId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Voiceid",
            "description": "Associated voice ID"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message",
            "description": "Status message or error"
          },
          "outputPath": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Outputpath",
            "description": "Path to generated audio file"
          },
          "createdAt": {
            "type": "string",
            "title": "Createdat",
            "description": "Job creation timestamp (ISO 8601)"
          },
          "startedAt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Startedat",
            "description": "Job start timestamp (ISO 8601)"
          },
          "finishedAt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finishedat",
            "description": "Job completion timestamp (ISO 8601)"
          },
          "requestParams": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Requestparams",
            "description": "Request parameters for generation"
          }
        },
        "type": "object",
        "required": [
          "id",
          "type",
          "status",
          "createdAt"
        ],
        "title": "Job",
        "description": "A generation job."
      },
      "ResolvedLine": {
        "properties": {
          "id": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Id",
            "description": "Line identifier"
          },
          "roleId": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Roleid",
            "description": "Role identifier"
          },
          "voiceId": {
            "type": "string",
            "minLength": 1,
            "title": "Voiceid",
            "description": "Resolved voice ID that will be used for this line"
          },
          "line": {
            "type": "string",
            "minLength": 1,
            "title": "Line",
            "description": "The text to speak"
          },
          "extra": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Extra",
            "description": "Performance hint"
          }
        },
        "type": "object",
        "required": [
          "id",
          "roleId",
          "voiceId",
          "line"
        ],
        "title": "ResolvedLine",
        "description": "A story line with resolved voice assignment.\n\nThis is the result of resolving a StoryLine using the resolution order:\nactorId \u2192 casting[roleId] \u2192 defaultVoiceId"
      },
      "Role": {
        "properties": {
          "roleId": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Roleid",
            "description": "Unique role identifier"
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "title": "Name",
            "description": "Role name"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes",
            "description": "Optional author notes"
          }
        },
        "type": "object",
        "required": [
          "roleId",
          "name"
        ],
        "title": "Role",
        "description": "A role in a story."
      },
      "StoryLine": {
        "properties": {
          "id": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Id",
            "description": "Unique line identifier"
          },
          "roleId": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Roleid",
            "description": "Role that speaks this line"
          },
          "line": {
            "type": "string",
            "minLength": 1,
            "title": "Line",
            "description": "The text to speak"
          },
          "extra": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Extra",
            "description": "Performance hint (e.g., 'curious, excited')"
          },
          "actorId": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Actorid",
            "description": "Optional per-line voice override. Highest priority in voice resolution: actorId \u2192 casting[roleId] \u2192 defaultVoiceId"
          }
        },
        "type": "object",
        "required": [
          "id",
          "roleId",
          "line"
        ],
        "title": "StoryLine",
        "description": "A single line in a story."
      },
      "StorySummary": {
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id",
            "description": "Unique story identifier (UUID)"
          },
          "slug": {
            "type": "string",
            "title": "Slug",
            "description": "URL-friendly identifier"
          },
          "title": {
            "type": "string",
            "title": "Title",
            "description": "Story title"
          }
        },
        "type": "object",
        "required": [
          "slug",
          "title"
        ],
        "title": "StorySummary",
        "description": "Summary of a story for list endpoints."
      },
      "StoryTemplate": {
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id",
            "description": "Unique story identifier (UUID). Set by server on creation."
          },
          "slug": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Slug",
            "description": "URL-friendly identifier derived from title. Set by server on creation."
          },
          "schemaVersion": {
            "type": "integer",
            "const": 1,
            "title": "Schemaversion",
            "description": "Schema version",
            "default": 1
          },
          "title": {
            "type": "string",
            "minLength": 1,
            "title": "Title",
            "description": "Story title"
          },
          "language": {
            "type": "string",
            "title": "Language",
            "description": "Language for TTS generation (e.g., 'English', 'Chinese', 'Spanish')",
            "default": "English"
          },
          "defaultVoiceId": {
            "type": "string",
            "minLength": 1,
            "title": "Defaultvoiceid",
            "description": "Fallback voice ID used when a role has no casting assignment and line has no actorId"
          },
          "roles": {
            "items": {
              "$ref": "#/components/schemas/Role"
            },
            "type": "array",
            "title": "Roles",
            "description": "Story roles"
          },
          "casting": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "string"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Casting",
            "description": "Optional map of roleId (as string key) to voiceId (string value). Example: {'0': 'narrator_male', '1': 'woman'}. Keys must be string representations of roleId integers.",
            "example": {
              "0": "narrator_male",
              "1": "woman",
              "2": "child"
            }
          },
          "lines": {
            "items": {
              "$ref": "#/components/schemas/StoryLine"
            },
            "type": "array",
            "title": "Lines",
            "description": "Story lines"
          }
        },
        "type": "object",
        "required": [
          "title",
          "defaultVoiceId",
          "roles",
          "lines"
        ],
        "title": "StoryTemplate",
        "description": "A story template with roles and lines.\n\nVoice assignment resolution order (for each line):\n1. line.actorId (if present) - per-line override\n2. casting[str(roleId)] (if present) - role-level assignment\n3. defaultVoiceId - fallback for all unassigned roles\n\nExample casting: {\"0\": \"narrator_male\", \"1\": \"woman\"} maps roleId 0 to narrator_male, roleId 1 to woman."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}